refactor: KidHome JSX 로직 함수로 분리
- 각 API fetching을 통해 변하는 JSX를 3개의 함수로 분리하여 추상화
- KidHome 컴포넌트의 코드 최소화, 가독성 개선
