+++ src/components/home/sumary/getKidSummaryContent.tsx
